requirements.txt
python-dotenv
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pgvector
openai
httpx
requests
PyJWT
